creating TeamRow
